Board position updating
The major addition here is board updating. `royalty reload` (which invokes `RoyaltyBoard.updateboard()`) will move players up if empty positions exist. In the future, this should be done automatically when there is a change. As far as I have tested, the `royalty` command is bug free, though I did not test manual changes to `board.yml` and I expect that will break it. We can look into automatically verifying and repairing the file later on.

I also did a lot of cleanup and am now using the LocalDateTime class to track time and write the time in `board.yml` as a readable date for manual editing. I've established the framework for the 30-day activity period but I've just now realized that I'm using `joined_time`. We'll need another variable to track how long a player has been offline.

I was having a lot of trouble syncing the memory of the board and the actual file but the proper way is to `RoyaltyBoard.save(boardFile)` after every change to ensure the changes persist when you `RoyaltyBoard.get()`.

Hope this longer message helps more than a set of bullet points. I've been committing a lot of changes at once so I figured this would be easier to understand everything.