isChallengeValid() method
Added the `Challenge` class for all things having to do with challenges. The `Challenge.isChallengeValid()` method takes an `attackingPlayerUUID` and `defendingPlayerUUID` to find whether a challenge would be valid or not. It has not been implemented into anything yet, but it's time to begin working on the framework for challenges. I chose to use **attacking** and **defending** instead of **challenging** and **challenged** to allow them to be easier to identify and differentiate.

I also added `RoyaltyBoard.getTeamNames()`, `RoyaltyBoard.getTribes()`, and `RoyaltyBoard.getValidPositions()` to centralize that permanent information.

The `last_online` variable now exists for every position on `board.yml` and is used for the offline timer. Both the `ListenJoin` and `ListenLeave` classes update this time, which will have no issues unless a player is connected for 30 days straight.