almost done with challenging
The challenge handshake system is beginning to near completion. Here is the process:
- Attacker initiates challenge
- Defender receives notification and either accepts or forfeits
- If accept, defender selects availability via web interface
- Upon completion, attacker must choose a time

That's as far as I've gotten. Next on the to-do list is to store dates of availability to appear in the notification.

Then comes the actual challenges.

I also fixed and optimized a bunch of various methods relating to Royalty Board management.

I've added Notification.create() which automatically attempts to send the notification and, if it fails, saves it to notification.yml. Super easy! At this point, notifications are sort of a set-it-and-forget-it thing.

I kind of glossed over it but Eye of Onyx runs an HTTP server for the time selection bc I do NOT want to figure out how to make an interface like that in Minecraft. It's also a lot easier for the user.

No other potential uses for the HTTP server right now, but since it's set up anyway, we should be on the lookout for uses.

We also have ONE entry in config.yml! It's simply the hostname used for the web address. I think in the future, we'll have timers and durations in here because there are a LOT of those in this plugin. That way, we aren't using hard-coded timers and they can be changed as we see fit.

With all the complexity of this plugin, I'm really glad it's all  under just a couple commands. I really do want to try to make it as hands-off as possible for both players and staff.